+++ src/strlcpy.c
@@ -55,5 +55,5 @@ local_strlcpy(char *dst, const char *src, size_t siz)
 			;
 	}
 
-	return(s - src - 1);	/* count does not include NUL */
+	return ((size_t)(s - src - 1));	/* count does not include NUL */
 }
